+++ pkg/apis/options/app.go
@@ -11,9 +11,10 @@ type Templates struct {
 	// If either file is missing, the default will be used instead.
 	Path string `flag:"custom-templates-dir" cfg:"custom_templates_dir"`
 
-	// CustomLogo is the path to a logo that should replace the default logo
+	// CustomLogo is the path or a URL to a logo that should replace the default logo
 	// on the sign_in page template.
 	// Supported formats are .svg, .png, .jpg and .jpeg.
+	// If URL is used the format support depends on the browser.
 	// To disable the default logo, set this value to "-".
 	CustomLogo string `flag:"custom-sign-in-logo" cfg:"custom_sign_in_logo"`
 
@@ -40,7 +41,7 @@ func templatesFlagSet() *pflag.FlagSet {
 	flagSet := pflag.NewFlagSet("templates", pflag.ExitOnError)
 
 	flagSet.String("custom-templates-dir", "", "path to custom html templates")
-	flagSet.String("custom-sign-in-logo", "", "path to an custom image for the sign_in page logo. Use \"-\" to disable default logo.")
+	flagSet.String("custom-sign-in-logo", "", "path or URL to an custom image for the sign_in page logo. Use \"-\" to disable default logo.")
 	flagSet.String("banner", "", "custom banner string. Use \"-\" to disable default banner.")
 	flagSet.String("footer", "", "custom footer string. Use \"-\" to disable default footer.")
 	flagSet.Bool("display-htpasswd-form", true, "display username / password login form if an htpasswd file is provided")