+++ pkg/app/pagewriter/sign_in_page_test.go
@@ -127,6 +127,11 @@ var _ = Describe("SignIn Page", func() {
 				expectedErr:  nil,
 				expectedData: "",
 			}),
+			Entry("with HTTPS URL", loadCustomLogoTableInput{
+				logoPath:     "https://raw.githubusercontent.com/oauth2-proxy/oauth2-proxy/master/docs/static/img/logos/OAuth2_Proxy_icon.png",
+				expectedErr:  nil,
+				expectedData: "<img src=\"https://raw.githubusercontent.com/oauth2-proxy/oauth2-proxy/master/docs/static/img/logos/OAuth2_Proxy_icon.png\" alt=\"Logo\" />",
+			}),
 			Entry("with an svg custom logo", loadCustomLogoTableInput{
 				logoPath:     "customDir/logo.svg",
 				expectedErr:  nil,